pyproject.toml
[project]
name = "tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imagehash>=4.3.2",
    "opencv-python>=4.12.0.88",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "sentence-transformers>=5.1.2",
    "torch",
    "torchaudio",
    "torchvision",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
]
